+++ catalog/entities/entities.go
@@ -70,6 +70,7 @@ type AreaToIngest struct {
 	AnnotationsURLs []string          `json:"annotations_urls"`
 	IsRetriable     bool              `json:"is_retriable"`
 	RetryCount      int               `json:"retry_count"`
+	StorageURI      string            `json:"storage_uri"` // If empty, use the default storage uri of the ingester
 }
 
 // GetConstellation returns the constellation from the user input